Added a command that ignores LNK4099 warnings
Added '/ignore:4099' at TheOneEditor's project properties in Linker > Command Line > Additional Options.

This ignores certain messages like the following:

5>AkMusicEngine.lib(AkMusicSwitchCntr.obj) : warning LNK4099: PDB 'AkMusicEngine.pdb' was not found with 'AkMusicEngine.lib(AkMusicSwitchCntr.obj)' or at 'D:\Users\lyl_j\Documents\GitHub\TheOneEngine\x64\TheOneBuilder\Debug\AkMusicEngine.pdb'; linking object as if no debug info

We hope this fastens the compile speed.
The reason why we are doing this it is since we are talking about Wwise .pdb files, we need them (download them again) and add them manually, but since we do not have an automatic installer, sh*t happens.

Co-Authored-By: Júlia Serra Trujillo <99959289+softdrawss@users.noreply.github.com>